refactor: Move recaf loader static methods to classpath util
